review comments, simplification
