+++ core/src/main/java/com/graphhopper/routing/util/DataFlagEncoder.java
@@ -118,10 +118,7 @@ public class DataFlagEncoder extends AbstractFlagEncoder {
 
     @Override
     public boolean acceptsTurnRelation(OSMTurnRelation relation) {
-        if (relation.isVehicleTypeConcernedByTurnRestriction(restrictions)) {
-            return true;
-        }
-        return false;
+        return relation.isVehicleTypeConcernedByTurnRestriction(restrictions);
     }
 
     @Override
