+++ core/src/test/java/com/graphhopper/routing/util/EncodingManagerTest.java
@@ -107,10 +107,7 @@ public class EncodingManagerTest {
 
             @Override
             public boolean acceptsTurnRelation(OSMTurnRelation relation) {
-                if (relation.isVehicleTypeConcernedByTurnRestriction(restrictions)) {
-                    return true;
-                }
-                return false;
+                return relation.isVehicleTypeConcernedByTurnRestriction(restrictions);
             }
 
             @Override
