+++ reader-osm/src/test/java/com/graphhopper/reader/osm/OSMTurnRelationTest.java
@@ -27,12 +27,17 @@ import com.graphhopper.storage.GraphHopperStorage;
 import com.graphhopper.util.EdgeExplorer;
 import org.junit.Test;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author Peter Karich
@@ -98,4 +103,35 @@ public class OSMTurnRelationTest {
         assertEquals(3, entry.nodeVia);
     }
 
+    @Test
+    public void testAcceptsTurnRelation() {
+        List<String> vehicleTypes = new ArrayList<>();
+        vehicleTypes.addAll(Arrays.asList("motorcar", "motor_vehicle", "vehicle"));
+        List<String> vehicleTypesExcept = new ArrayList<>();
+        OSMTurnRelation osmTurnRelation = new OSMTurnRelation(1, 1, 1, OSMTurnRelation.Type.NOT);
+        assertTrue(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+
+        vehicleTypesExcept.add("bus");
+        osmTurnRelation.setVehicleTypesExcept(vehicleTypesExcept);
+        assertTrue(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+
+        vehicleTypesExcept.clear();
+        vehicleTypesExcept.add("vehicle");
+        osmTurnRelation.setVehicleTypesExcept(vehicleTypesExcept);
+        assertFalse(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+
+        vehicleTypesExcept.clear();
+        vehicleTypesExcept.add("motor_vehicle");
+        vehicleTypesExcept.add("vehicle");
+        osmTurnRelation.setVehicleTypesExcept(vehicleTypesExcept);
+        assertFalse(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+
+        vehicleTypesExcept.clear();
+        osmTurnRelation.setVehicleTypeRestricted("bus");
+        osmTurnRelation.setVehicleTypesExcept(vehicleTypesExcept);
+        assertFalse(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+
+        osmTurnRelation.setVehicleTypeRestricted("vehicle");
+        assertTrue(osmTurnRelation.isVehicleTypeConcernedByTurnRestriction(vehicleTypes));
+    }
 }
